refactor(grovedb): split aggregate_count proof module into submodules
The single 750-line `aggregate_count.rs` had grown to cover four
distinct concerns: public API, classification, the leaf-only chain
walker, and the per-key carrier walker (plus a pile of shared
helpers). Split into a directory module with one file per concern.

```
grovedb/src/operations/proof/aggregate_count/
├── mod.rs            227 lines — public API (impl GroveDb,
│                                  verify_aggregate_count_query +
│                                  verify_aggregate_count_query_per_key),
│                                  require_v1_envelope, submodule wiring
├── classification.rs  77 lines — AggregateCountClassification struct +
│                                  classify_aggregate_count_path_query
├── leaf_chain.rs      77 lines — verify_v1_leaf_chain (legacy
│                                  single-u64 walker)
├── per_key.rs        221 lines — per-key carrier walker
│                                  (with_classification, _per_key,
│                                  _carrier_layer, _subquery_path)
└── helpers.rs        270 lines — shared utilities (decode envelope,
                                  verify_count_leaf, expect_merk_bytes,
                                  verify_single_key_layer_proof_v0,
                                  OuterMatch + execute_carrier_layer_proof,
                                  enforce_lower_chain)
```

Visibility is `pub(super)` everywhere — the submodules are crate
implementation detail, only `verify_aggregate_count_query` and
`verify_aggregate_count_query_per_key` are public.

Pure code reorganization — no behavior change. All 62 grovedb +
28 grovedb-query aggregate-count tests still pass; workspace +
clippy clean.
